#31 Creates request classes for stat api (#54)
