+++ src/main/java/mobi/appcent/helium/api/IStatApi.java
@@ -1,12 +1,15 @@
 package mobi.appcent.helium.api;
 
+import mobi.appcent.helium.request.stat.BlockchainStatsRequest;
+import mobi.appcent.helium.request.stat.TokenSupplyRequest;
+
 import java.io.IOException;
 
 /**
  * Created by erenalpaslan on 28.10.2022
  */
 public interface IStatApi {
-    public StatApi.APIgetBlockchainStatsRequest getBlockchainStats() throws IOException;
+    BlockchainStatsRequest getBlockchainStats() throws IOException;
 
-    public StatApi.APIgetTokenSupplyRequest getTokenSupply() throws IOException;
+    TokenSupplyRequest getTokenSupply() throws IOException;
 }
